feat: use nginx for prod frontend server
The frontend Dockerfile has been modified to include two stages, one for
building and another with nginx to serve the files.

As a result, two new docker-compose files have been added:

    1. docker-compose.override.yml
    2. docker-compose.prod.yml

docker-compose.override.yml is for development specific settings. The
`target` option under `build` is set to only build the build stage
without nginx. There are volumes for the backend, frontend and nginx
services to reload changes without having to rebuild. Additionally,
ports are exposed for direct access to each service.

docker-compose.prod.yml is configured to use the `serve` stage for the
frontend.